requirements.txt
# Core Flask dependencies
Flask==3.0.0
Flask-CORS==4.0.0

# Twilio SDK for voice calls
twilio==8.10.0

# Google Generative AI (Gemini)
google-generativeai==0.3.2

# Environment and configuration
python-dotenv==1.0.0

# Database (SQLite is built into Python)
# No additional dependencies needed for SQLite

# Development dependencies
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0

# Optional: For better development experience
watchdog==3.0.0  # For file watching during development
colorama==0.4.6   # For colored terminal output

# Production dependencies (optional)
gunicorn==21.2.0  # WSGI server for production